requirements.txt
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
diffusers==0.29.0
transformers==4.39.3
accelerate==0.27.0
safetensors==0.4.2
runpod==1.3.0
Pillow==10.3.0
xformers==0.0.25
